1.2.0
New feature :

Allow to use only imagex.txt and imagey.txt to create a 3D object. Using imagez.txt without using the two others might get an issue. I will try to fix it if it appears, add it to the To Do list.

Zoom in the 3D object by pressing the key n and zoom out by pressing the key b. Replace keys to mouse scrolling could be nice. Added to the To Do list.

Minimize the number of point by using the key o, that feature will remove point inside the 3D object and then allow to move it faster. This feature might take some time to execute.

Save the 3D object as a .3DPiaf by using the key s. So after Minimize the number of point, the 3D object can be saved to avoid to minimize it each time it's loaded.

Load the 3D object by using the key l.